requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
PyYAML>=6.0
scipy>=1.7.0
tqdm>=4.62.0
opencv-python>=4.5.0
scikit-image>=0.19.0
matplotlib>=3.5.0
tensorboard>=2.8.0
blobfile>=2.0.0
mpi4py>=3.1.0
